A. CheckOverlaping/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
begins==0.9
colorama==0.4.1
more-itertools==7.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
scandir==1.10.0
six==1.12.0
